优化主要部分的代码

+++ app/src/main/java/com/android/reverse/collecter/DexFileInfo.java
@@ -31,7 +31,7 @@ public class DexFileInfo {
         return mCookie;
     }
 
-    public void setmCookie(int mCookie) {
+    public void setmCookie(long mCookie) {
         this.mCookie = mCookie;
     }
